refactor(*): reorder files on packages
