+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
 
 	<application
 		android:label="@string/app_name"
-		android:name="org.flyve.inventory.agent.InventoryAgentApp"
+		android:name="org.flyve.inventory.agent.ui.InventoryAgentApp"
 		android:vmSafeMode="true"
 		android:theme="@style/AppTheme"
 		android:icon="@drawable/icon">
@@ -44,7 +44,7 @@
 			android:name="com.bugsnag.android.API_KEY"
 			android:value="b6bc7ce0da33f92fe6c49c568af43963"/>
 
-		<activity android:name=".SplashActivity"
+		<activity android:name=".ui.ActivitySplash"
 			android:theme="@style/NoActionBar">
 			<intent-filter>
 				<action
@@ -54,20 +54,20 @@
 			</intent-filter>
 		</activity>
 		<activity
-			android:name="org.flyve.inventory.agent.MainActivity"
+			android:name="org.flyve.inventory.agent.ui.ActivityMain"
 			android:label="@string/app_name"
 			>
 		</activity>
 		<activity
-			android:name="org.flyve.inventory.agent.InventoryParametersPreference"
+			android:name="org.flyve.inventory.agent.preference.InventoryParametersPreference"
 			android:label="@string/app_name">
 		</activity>
 		<activity
-			android:name="org.flyve.inventory.agent.GlobalParametersPreference"
+			android:name="org.flyve.inventory.agent.preference.GlobalParametersPreference"
 			android:label="@string/app_name">
 		</activity>
 		<activity
-			android:name="org.flyve.inventory.agent.InventoryActivity"
+			android:name="org.flyve.inventory.agent.ui.ActivityInventory"
 			android:label="@string/app_name"
 			android:launchMode="standard"
 			android:theme="@style/NoActionBar"
@@ -79,19 +79,19 @@
 			android:label="Inventory Agent"
 			android:description="@string/agent_description"
 			android:icon="@drawable/icon"
-			android:name="org.flyve.inventory.agent.InventoryService"
+			android:name="org.flyve.inventory.agent.service.InventoryService"
 			tools:ignore="ExportedService">
 			<intent-filter>
 				<action android:name="org.flyve.inventory.service" />
 			</intent-filter>
 		</service>
-		<receiver android:name="org.flyve.inventory.agent.BootStartAgent">
+		<receiver android:name="org.flyve.inventory.agent.broadcast.BootStartAgent">
 			<intent-filter>
 				<action android:name="android.intent.action.BOOT_COMPLETED" />
 			</intent-filter>
 		</receiver>
 
-		<receiver android:name="org.flyve.inventory.agent.TimeAlarm">
+		<receiver android:name="org.flyve.inventory.agent.broadcast.TimeAlarm">
 			<intent-filter>
 				<action android:name="org.flyve.inventory.agent.ALARM"/>
 			</intent-filter>
